adapter/pyproject.toml
[project]
name = "retriva-adapter"
version = "0.1.0"
description = "Thin adapter (Pattern B-1) — mirrors Open WebUI file uploads into Retriva ingestion"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "httpx>=0.28,<1",
    "pydantic-settings>=2.7,<3",
    "aiosqlite>=0.21,<1",
    "apscheduler>=3.11,<4",
    "structlog>=25.1,<26",
    "prometheus-client>=0.22,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.25,<1",
    "respx>=0.22,<1",
    "coverage>=7,<8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adapter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
